+++ lab1.1/experi_mm/src/matmult.h
@@ -1,3 +1,3 @@
 #define N 16 
 
-void matmult(short int A[N][N],short int B[N][N],int C[N][N]);
+void matmult(short A[N][N], short B[N][N], short C[N][N]);
